update help and guide
